+++ lib/merchant_sidekick.rb
@@ -36,6 +36,8 @@ require 'merchant_sidekick/gateway'
 require 'merchant_sidekick/payment'
 
 require 'merchant_sidekick/active_merchant/credit_card_payment'
+
+require 'merchant_sidekick/active_merchant/gateways/base'
 require 'merchant_sidekick/active_merchant/gateways/authorize_net_gateway'
 require 'merchant_sidekick/active_merchant/gateways/paypal_gateway'