+++ Changelog.md
@@ -10,6 +10,7 @@
 * Removed option to load gateway configuration from database
 * Added gateway spec
 * Allow for MerchantSidekick::Gateway.default_gateway to use type name
+* Refactored default_gateway for active merchant gateway types into base class
 
 ## 0.4.0 (2011-12-27)